+++ src/EVM/Types.hs
@@ -554,6 +554,7 @@ data EvmError
   | MaxInitCodeSizeExceeded W256 (Expr EWord)
   | InvalidFormat
   | PrecompileFailure
+  | NonexistentPrecompile Addr
   | ReturnDataOutOfBounds
   | NonceOverflow
   | BadCheatCode String FunctionSelector
@@ -591,6 +592,7 @@ data PartialExec
   = UnexpectedSymbolicArg { pc :: Int, addr :: Expr EAddr, opcode :: String, msg  :: String, args  :: [SomeExpr] }
   | MaxIterationsReached  { pc :: Int, addr :: Expr EAddr }
   | JumpIntoSymbolicCode  { pc :: Int, addr :: Expr EAddr, jumpDst :: Int }
+  | PrecompileMissing     { pc :: Int, addr :: Expr EAddr, preAddr :: Addr }
   | CheatCodeMissing      { pc :: Int, addr :: Expr EAddr, selector :: FunctionSelector }
   | BranchTooDeep         { pc :: Int, addr :: Expr EAddr}
   deriving (Show, Eq, Ord)