+++ test/test.hs
@@ -71,7 +71,7 @@ import EVM.Solvers
 import EVM.Stepper qualified as Stepper
 import EVM.SymExec
 import EVM.Test.FuzzSymExec qualified as FuzzSymExec
-import EVM.Test.Utils (runSolidityTest, runSolidityTestCustom)
+import EVM.Test.Utils (runForgeTest, runForgeTestCustom)
 import EVM.Traversals
 import EVM.Types hiding (Env)
 import EVM.Effects
@@ -662,6 +662,12 @@ tests = testGroup "hevm"
         e2 = ConcreteBuf "Definitely not the same!"
       equal <- checkEquiv e1 e2
       assertBoolM "Should not be equivalent!" $ not equal
+    , testNoSimplify "simplify-comparison-GEq" $ do
+      let
+        expr = PEq (Lit 0x1) (GEq (Var "v") (Lit 0x2))
+        simp = Expr.simplifyProp expr
+      ret <- checkEquivPropAndLHS expr simp
+      assertEqualM "Must be equivalent" True ret
     ]
   -- These tests fuzz the simplifier by generating a random expression,
   -- applying some simplification rules, and then using the smt encoding to
@@ -691,7 +697,7 @@ tests = testGroup "hevm"
     , testProperty "byte-simplification" $ \(expr :: Expr Byte) -> propNoSimp $ do
         let simplified = Expr.simplify expr
         checkEquivAndLHS expr simplified
-    , testProperty "word-simplification" $ \(ZeroDepthWord expr) -> propNoSimp $ do
+    , askOption $ \(QuickCheckTests n) -> testProperty "word-simplification" $ withMaxSuccess (min n 20) $ \(ZeroDepthWord expr) -> propNoSimp $ do
         let simplified = Expr.simplify expr
         checkEquivAndLHS expr simplified
     , testProperty "readStorage-equivalance" $ \(store, slot) -> propNoSimp $ do
@@ -1052,6 +1058,13 @@ tests = testGroup "hevm"
         case runGetOrFail (getAbi (abiValueType x)) (runPut (putAbi x)) of
           Right ("", _, x') -> x' == x
           _ -> False
+    , test "ABI-negative-small-int" $ do
+        let bs = hex "ffffd6" -- -42 as int24
+        let padded = BS.replicate (32 - BS.length bs) 0 <> bs -- padded to 32 bytes
+        let withSelector = BS.replicate 4 0 <> padded -- added extra 4 bytes, simulating selector
+        case decodeAbiValues [AbiIntType 24] withSelector of
+          [AbiInt 24 val] -> assertEqualM "Incorrectly decoded int24 value" (-42) val
+          _ -> internalError "Error in decoding function"
     ]
   , testGroup "Solidity-Expressions"
     [ test "Trivial" $
@@ -1681,8 +1694,8 @@ tests = testGroup "hevm"
                 }
               }
             |]
-        Right e <- reachableUserAsserts c Nothing
-        assertBoolM "The expression is not partial" $ Expr.containsNode isPartial e
+        r <- allBranchesFail c Nothing
+        assertBoolM "all branches must fail" (isRight r)
       ,
       -- TODO: we can't deal with symbolic jump conditions
       expectFail $ test "call-zero-inited-var-thats-a-function" $ do
@@ -1970,7 +1983,7 @@ tests = testGroup "hevm"
   , testGroup "Dapp-Tests"
     [ test "Trivial-Pass" $ do
         let testFile = "test/contracts/pass/trivial.sol"
-        runSolidityTest testFile ".*" >>= assertEqualM "test result" (True, True)
+        runForgeTest testFile ".*" >>= assertEqualM "test result" (True, True)
     , test "Foundry" $ do
         -- quick smokecheck to make sure that we can parse ForgeStdLib style build outputs
         -- return is a pair of (No Cex, No Warnings)
@@ -2002,59 +2015,59 @@ tests = testGroup "hevm"
               , ("test/contracts/fail/symbolicFail.sol",      "prove_symb_fail_allrev_selector.*", (False, False))
               , ("test/contracts/fail/symbolicFail.sol",      "prove_symb_fail_somerev_selector.*", (False, True))]
         forM_ cases $ \(testFile, match, expected) -> do
-          actual <- runSolidityTestCustom testFile match Nothing Nothing False mempty Foundry
+          actual <- runForgeTestCustom testFile match Nothing Nothing False mempty
           putStrLnM $ "Test result for " <> testFile <> " match: " <> T.unpack match <> ": " <> show actual
           assertEqualM "Must match" expected actual
     , test "Trivial-Fail" $ do
         let testFile = "test/contracts/fail/trivial.sol"
-        runSolidityTest testFile "prove_false" >>= assertEqualM "test result" (False, False)
+        runForgeTest testFile "prove_false" >>= assertEqualM "test result" (False, False)
     , test "Abstract" $ do
         let testFile = "test/contracts/pass/abstract.sol"
-        runSolidityTest testFile ".*" >>= assertEqualM "test result" (True, True)
+        runForgeTest testFile ".*" >>= assertEqualM "test result" (True, True)
     , test "Constantinople" $ do
         let testFile = "test/contracts/pass/constantinople.sol"
-        runSolidityTest testFile ".*" >>= assertEqualM "test result" (True, True)
+        runForgeTest testFile ".*" >>= assertEqualM "test result" (True, True)
     , test "ConstantinopleMin" $ do
         let testFile = "test/contracts/pass/constantinople_min.sol"
-        runSolidityTest testFile ".*" >>= assertEqualM "test result" (True, True)
+        runForgeTest testFile ".*" >>= assertEqualM "test result" (True, True)
     , test "Prove-Tests-Pass" $ do
         let testFile = "test/contracts/pass/dsProvePass.sol"
-        runSolidityTest testFile ".*" >>= assertEqualM "test result" (True, True)
+        runForgeTest testFile ".*" >>= assertEqualM "test result" (True, True)
     , test "prefix-check-for-dapp" $ do
         let testFile = "test/contracts/fail/check-prefix.sol"
-        runSolidityTest testFile "prove_trivial" >>= assertEqualM "test result" (False, False)
+        runForgeTest testFile "prove_trivial" >>= assertEqualM "test result" (False, False)
     , test "transfer-dapp" $ do
         let testFile = "test/contracts/pass/transfer.sol"
-        runSolidityTest testFile "prove_transfer" >>= assertEqualM "should prove transfer" (True, True)
+        runForgeTest testFile "prove_transfer" >>= assertEqualM "should prove transfer" (True, True)
     , test "nonce-issues" $ do
         let testFile = "test/contracts/pass/nonce-issues.sol"
-        runSolidityTest testFile "prove_prank_addr_exists" >>= assertEqualM "should not bail" (True, True)
-        runSolidityTest testFile "prove_nonce_addr_nonexistent" >>= assertEqualM "should not bail" (True, True)
+        runForgeTest testFile "prove_prank_addr_exists" >>= assertEqualM "should not bail" (True, True)
+        runForgeTest testFile "prove_nonce_addr_nonexistent" >>= assertEqualM "should not bail" (True, True)
     , test "Prove-Tests-Fail" $ do
         let testFile = "test/contracts/fail/dsProveFail.sol"
-        runSolidityTest testFile "prove_trivial" >>= assertEqualM "prove_trivial" (False, False)
-        runSolidityTest testFile "prove_trivial_dstest" >>= assertEqualM "prove_trivial_dstest" (False, False)
-        runSolidityTest testFile "prove_add" >>= assertEqualM "prove_add" (False, True)
-        runSolidityTestCustom testFile "prove_smtTimeout" (Just 1) Nothing False mempty Foundry
+        runForgeTest testFile "prove_trivial" >>= assertEqualM "prove_trivial" (False, False)
+        runForgeTest testFile "prove_trivial_dstest" >>= assertEqualM "prove_trivial_dstest" (False, False)
+        runForgeTest testFile "prove_add" >>= assertEqualM "prove_add" (False, True)
+        runForgeTestCustom testFile "prove_smtTimeout" (Just 1) Nothing False mempty
           >>= assertEqualM "prove_smtTimeout" (True, False)
-        runSolidityTest testFile "prove_multi" >>= assertEqualM "prove_multi" (False, True)
-        runSolidityTest testFile "prove_distributivity" >>= assertEqualM "prove_distributivity" (False, True)
+        runForgeTest testFile "prove_multi" >>= assertEqualM "prove_multi" (False, True)
+        runForgeTest testFile "prove_distributivity" >>= assertEqualM "prove_distributivity" (False, True)
     , test "Loop-Tests" $ do
         let testFile = "test/contracts/pass/loops.sol"
-        runSolidityTestCustom testFile "prove_loop" Nothing (Just 10) False mempty Foundry  >>= assertEqualM "test result" (True, False)
-        runSolidityTestCustom testFile "prove_loop" Nothing (Just 100) False mempty Foundry >>= assertEqualM "test result" (False, False)
+        runForgeTestCustom testFile "prove_loop" Nothing (Just 10) False mempty  >>= assertEqualM "test result" (True, False)
+        runForgeTestCustom testFile "prove_loop" Nothing (Just 100) False mempty >>= assertEqualM "test result" (False, False)
     , test "Cheat-Codes-Pass" $ do
         let testFile = "test/contracts/pass/cheatCodes.sol"
-        runSolidityTest testFile ".*" >>= assertEqualM "test result" (True, False)
+        runForgeTest testFile ".*" >>= assertEqualM "test result" (True, False)
     , test "Cheat-Codes-Fork-Pass" $ do
         let testFile = "test/contracts/pass/cheatCodesFork.sol"
-        runSolidityTest testFile ".*" >>= assertEqualM "test result" (True, True)
+        runForgeTest testFile ".*" >>= assertEqualM "test result" (True, True)
     , test "Unwind" $ do
         let testFile = "test/contracts/pass/unwind.sol"
-        runSolidityTest testFile ".*" >>= assertEqualM "test result" (True, True)
+        runForgeTest testFile ".*" >>= assertEqualM "test result" (True, True)
     , test "Keccak" $ do
         let testFile = "test/contracts/pass/keccak.sol"
-        runSolidityTest testFile "prove_access" >>= assertEqualM "test result" (True, True)
+        runForgeTest testFile "prove_access" >>= assertEqualM "test result" (True, True)
     ]
   , testGroup "max-iterations"
     [ test "concrete-loops-reached" $ do
@@ -4858,7 +4871,7 @@ tests = testGroup "hevm"
             _     -> liftIO $ assertFailure "Must have exactly one cex"
           let def = fromRight (error "cannot be") $ defaultSymbolicValues $ subModel cex (AbstractBuf "txdata")
           let buf = prettyBuf $ Expr.concKeccakSimpExpr def
-          assertBoolM "Must start with specific string" (T.isPrefixOf "0xfffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffe0cf" buf)
+          assertBoolM "Must start with specific string" (T.isPrefixOf "0xfffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffe0c" buf)
       , test "eq-yul-simple-cex" $ do
         Right aPrgm <- liftIO $ yul ""
           [i|
@@ -5571,15 +5584,18 @@ checkEquivAndLHS orig simp = do
 
 checkEquivBase :: (Eq a, App m) => (a -> a -> Prop) -> a -> a -> Bool -> m (Maybe Bool)
 checkEquivBase mkprop l r expect = do
-  withSolvers Z3 1 1 (Just 1) $ \solvers -> do
-     res <- checkSatWithProps solvers [mkprop l r]
-     let ret = case res of
-           Qed -> Just True
-           Cex {} -> Just False
-           Error _ -> Just (not expect)
-           Unknown _ -> Nothing
-     when (ret == Just (not expect)) $ liftIO $ print res
-     pure ret
+  config <- readConfig
+  let noSimplifyEnv = Env {config = config {simp = False}}
+  liftIO $ runEnv noSimplifyEnv $ do
+    withSolvers Z3 1 1 (Just 1) $ \solvers -> do
+      res <- checkSatWithProps solvers [mkprop l r]
+      let ret = case res of
+            Qed -> Just True
+            Cex {} -> Just False
+            Error _ -> Just (not expect)
+            Unknown _ -> Nothing
+      when (ret == Just (not expect)) $ liftIO $ print res
+      pure ret
 
 -- | Takes a runtime code and calls it with the provided calldata
 
@@ -6253,7 +6269,7 @@ genBuf bound sz = oneof
 genStorage :: Int -> Gen (Expr Storage)
 genStorage 0 = oneof
   [ liftM2 AbstractStore arbitrary (pure Nothing)
-  , fmap ConcreteStore arbitrary
+  , fmap ConcreteStore $ resize 5 arbitrary
   ]
 genStorage sz = liftM3 SStore key val subStore
   where